refactor(asmdb): disentangle decodetree augmentation from antd presentation
